refactor: 유효성 검사를 위한 당첨번호 입력 구문 수정
